[team_info] refactor: 파일명 변경에 따른 import 경로 수정
